+++ tcex/api/tc/v3/case_attributes/case_attribute_model.py
@@ -104,7 +104,6 @@ class CaseAttributeModel(
         None,
         description='The attribute value.',
         methods=['POST', 'PUT'],
-        min_length=1,
         read_only=False,
         title='value',
     )