+++ tcex/api/tc/v3/victim_attributes/victim_attribute_model.py
@@ -97,7 +97,6 @@ class VictimAttributeModel(
         None,
         description='The attribute value.',
         methods=['POST', 'PUT'],
-        min_length=1,
         read_only=False,
         title='value',
     )